requirements.txt
virustotal-api>=1.1.11
watchdog>=2.1.3
requests>=2.25.1